multi-crop-disease-detection/requirements.txt
FlFlask==2.3.3
tensorflow==2.13.0
numpy==1.24.3
pillow==10.0.0
opencv-python==4.8.0.74
python-dotenv==1.0.0
email-validator==2.0.0
flask-cors==4.0.0
requests==2.31.0
matplotlib==3.7.2
scikit-learn==1.3.0
pandas==2.0.3